[stable/20.12] Fix ovn-northd certificate reload workaround
The charm currently tries to reload the ovn-northd service, but
that is unfortunately not implemented by the init/systemd script.

Restart the service instead.

Note that this is a stable/20.03 and stable/20.12-only fix as this was fixed in OVN upstream at 21.06 [0].

0: https://github.com/ovn-org/ovn/commit/8de4f8005f210
Closes-Bug: #1895303
Change-Id: I50f2e76a42f7d8f305dda216048c6c5ec62a6c0e
(cherry picked from commit 87fe7cc6aefce1bc0d4dc875c8dc1928bf65af6c)