[WIP] Default to OVN 22.03 for fresh deployments on Focal
__NOTE__
This PR is a cherry-pick of in-flight change on main
branch and also references in-flight stable backport for
charm-helpers.

Do not merge until these are merged.
__NOTE__

Add `ovn-source` configuration option which should be used in
addition to the `source` configuration option.

The default of the `ovn-source` configuration option is
determined at runtime based on whether the end user has provided
config, whether its a fresh or upgraded charm deployment and which
series it is deployed on.

Func-Test-Pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/927
Change-Id: I54f2339a16bd958c29f877394209039f345e106e