Build on 23.04 base to run on 23.04 (Lunar)
Currently the charm is built on 22.04 which has Python 3.10, so when
building the python binary wheels they are associated to Python 3.10,
this makes them unsable on Ubuntu 23.04 which has Python 3.11.

This change makes the building process for Lunar as a separate job.

Closes-Bug: #2019996
Change-Id: I960fe72a1f6a2f7864826c6372c355119157e8b2
(cherry picked from commit 35509bee45a12f78d0b1440cfa7ebe8ab142ddf2)