Ensure juju displays errmsg from SSL issues
There are times when the certificate expires, and the charm goes into error
state. This change ensures the right message is displayed, and the right actions
can be taken. In errored state the vault reissue action doesn't work, this
update would allow for the certs to be updated as the charm is not in error
state.

Func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1332

Signed-off-by: Arif Ali <arif.ali@canonical.com>